refactor(cli): factor serve into ServeConfig (dashboard step 0)
Extracts the ~240-line Commands::Serve body and run_minimal_serve from
main.rs into a new crates/ctxd-cli/src/serve.rs with a typed config
struct and a single async fn entry point. Pure refactor — `ctxd serve`
behavior, ordering, and side effects are preserved.

Why: the upcoming `ctxd dashboard` subcommand needs the same bootstrap
(open store, build cap engine, mount HTTP admin, optional federation /
wire / MCP) but with most transports off. Without this factoring, the
dashboard handler would copy-paste the bootstrap, drift over time, and
risk diverging from `serve` on every change.

ServeConfig adds two fields beyond the existing CLI flags:
- wire_bind: Option<String> (None = skip wire transport)
- federation: bool (false = skip peer manager + replication tasks)

Today's `ctxd serve` always sets both ON, so behavior is unchanged.
The dashboard subcommand (step 8) will set both OFF.

Verification: cargo test --workspace passes (50+ test result lines, no
failures); manual smoke of `ctxd serve --bind 127.0.0.1:17777
--wire-bind 127.0.0.1:17778` shows HTTP admin (200 on /v1/stats and
/health), wire protocol listener, federation identity bootstrapped, MCP
stdio ready — identical to pre-refactor.
