+++ crates/ctxd-cli/src/lib.rs
@@ -11,4 +11,5 @@ pub mod federation;
 pub mod protocol;
 pub mod query;
 pub mod rate_limit;
+pub mod serve;
 pub mod storage_selector;
